Apply SRP in the Layouts package
