+++ TagsCloudContainer/Visualization/Measurers/IWordMeasurer.cs
@@ -1,0 +1,10 @@
+﻿using System.Drawing;
+using TagsCloudContainer.Data;
+
+namespace TagsCloudContainer.Visualization.Layouts
+{
+    public interface IWordMeasurer
+    {
+        (Font font, Size size) Measure(Word word);
+    }
+}
